Prisma4 jest pnpm (#284)
* start new project

* github action to test

* merge main

* fix actions

* fix actions

* build before test

* ignor pnpm-lock.yaml from formatting

* max_old_space_size

* fix snap

* update prisma version

* update dmmf snapshots

+++ packages/generator/src/admin/createFile.ts
@@ -3,6 +3,5 @@ import { join } from 'path';
 
 export const createFile = (path: string, fileName: string, content: string) => {
   !existsSync(path) && mkdirSync(path, { recursive: true });
-  !existsSync(join(path, fileName)) &&
-    writeFileSync(join(path, fileName), content);
+  !existsSync(join(path, fileName)) && writeFileSync(join(path, fileName), content);
 };
